requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
websockets==13.0.1
google-genai==0.3.0
python-multipart==0.0.10
python-dotenv==1.0.1
